Commands: Remove EvaluateCsharp
This module is useless in a sense, the containers that hold this bot are only permitted up to a GiB of memory, and when getting influx from 30k+ guilds, the average memory usage stays around 810-900MiB, this command, when executed will immediately allocate nearly around 150-200MiB of memory, causing the container to crash due to OOM almost instantly, and even when forcing a GC, this memory is not deallocated until some time after.

After some thoughts, this command has been removed because it impedes on performance.